DRAFT - Attempt to improve compatibility with Webkit GTK
Deadlock / Freeze when Webkit calls Java which calls Webkit Javascript
evaluation. Webkit GKT seems not abel to handle such subexecutions and
is waiting for the initial javascript evaluation result.

Idea here is to move the Java code which needs to evaluation some
javascript and get the result to a runnable which is asynchronously
triggered.